fix: delete aelf-design and optimize demo structure and update readme
