+++ packages/bridge/package.json
@@ -28,7 +28,6 @@
     "@aelf-web-login/wallet-adapter-base": "workspace:*",
     "@portkey/did-ui-react": "^2.6.2",
     "@reduxjs/toolkit": "^2.2.3",
-    "aelf-design": "1.0.0-alpha.17",
     "react": "^18.0.0",
     "react-dom": "^18.0.0"
   },
