+++ packages/starter/doc-site/src/Demo/LoginDemo.tsx
@@ -1,0 +1,101 @@
+import { useState } from 'react';
+import { Button, message, Divider, Flex } from 'antd';
+import { useConnectWallet } from '@aelf-web-login/wallet-adapter-react';
+
+const LoginDemo = () => {
+  const {
+    connectWallet,
+    disConnectWallet,
+    walletInfo,
+    lock,
+    isLocking,
+    walletType,
+    isConnected,
+    getAccountByChainId,
+    getWalletSyncIsCompleted,
+  } = useConnectWallet();
+  console.log('LoginDemo init----------');
+  const [aelfAccount, setAelfAccount] = useState<string>('');
+  const [tdvwAccount, setTdvwAccount] = useState<string>('');
+  const [syncIsCompleted, setSyncIsCompleted] = useState<string | boolean>(false);
+  const [syncIsCompletedTDVW, setSyncIsCompletedTDVW] = useState<string | boolean>(false);
+
+  const onConnectBtnClickHandler = async () => {
+    try {
+      const rs = await connectWallet();
+      console.log('rs', rs);
+    } catch (e: any) {
+      message.error(e.message);
+    }
+  };
+
+  const onDisConnectBtnClickHandler = () => {
+    disConnectWallet();
+  };
+
+  const onGetAccountByAELFHandler = async () => {
+    const account = await getAccountByChainId('AELF');
+    setAelfAccount(account);
+  };
+
+  const onGetAccountBytDVWHandler = async () => {
+    const account = await getAccountByChainId('tDVW');
+    setTdvwAccount(account);
+  };
+
+  const onGetWalletSyncIsCompletedHandler = async () => {
+    const account = await getWalletSyncIsCompleted('AELF');
+    setSyncIsCompleted(account);
+  };
+
+  const onGetWalletSyncIsCompletetDVWdHandler = async () => {
+    const account = await getWalletSyncIsCompleted('tDVW');
+    setSyncIsCompletedTDVW(account);
+  };
+
+  return (
+    <div>
+      <Flex gap="small">
+        <Button type="primary" onClick={onConnectBtnClickHandler} disabled={isConnected}>
+          {isLocking ? 'unlock' : 'connect'}
+        </Button>
+        <Button type="primary" onClick={lock} disabled={!walletInfo}>
+          lock
+        </Button>
+        <Button type="primary" onClick={onDisConnectBtnClickHandler} disabled={!isConnected}>
+          disConnect
+        </Button>
+      </Flex>
+      <div>
+        walletInfo:
+        <pre style={{ overflow: 'auto', height: '300px' }}>
+          {JSON.stringify(walletInfo, null, 4)}
+        </pre>
+      </div>
+      <div>walletType:{walletType}</div>
+      <Divider />
+      <Button type="primary" onClick={onGetAccountByAELFHandler}>
+        getAccountByChainId-AELF
+      </Button>
+      <div>getAccountByChainId-AELF:{aelfAccount}</div>
+
+      <Button type="primary" onClick={onGetAccountBytDVWHandler}>
+        getAccountByChainId-tDVW
+      </Button>
+      <div>getAccountByChainId-tDVW:{tdvwAccount}</div>
+      <Divider />
+      <Button type="primary" onClick={onGetWalletSyncIsCompletedHandler}>
+        getWalletSyncIsCompleted-AELF
+      </Button>
+      <div>getWalletSyncIsCompleted-AELF:{syncIsCompleted}</div>
+
+      <Button type="primary" onClick={onGetWalletSyncIsCompletetDVWdHandler}>
+        getWalletSyncIsCompleted-tDVW
+      </Button>
+      <div>getWalletSyncIsCompleted-tDVW:{syncIsCompletedTDVW}</div>
+      <Divider />
+    </div>
+  );
+};
+
+export default LoginDemo;
